대댓글 작성시 쿼리스트링 이동 해결

+++ src/main/java/hello/login/web/model/CommentDTO.java
@@ -8,9 +8,9 @@ import lombok.Setter;
 @Setter
 public class CommentDTO extends CommonDTO{
 
-    private Long idx;
+    private Integer idx;
 
-    private Long boardIdx;
+    private Integer boardIdx;
 
     private String content;
